garmin flag scss file moved into product carousel folder
